refactor levelgraph, split in mutlipe classes for more readabillity
